Deal with fixture init failures in boost.test
On sufficiently fatal errors (such as an exception thrown from a global fixture ctor) boost.test
ignores --output_format=XML and dumps the error message as is. Check for those cases explicitly
instead of crashing due to malformed/empty xml.

(alternatively we could check for result code 200 (aka boost::exit_exception_failure), but I'm not
sure that boost.test doesn't return it in other cases)

+++ pytest_cpp/boost.py
@@ -69,6 +69,13 @@ class BoostTestFacade(object):
                                     returncode=p.returncode))
             return [failure]
 
+        if report is not None and (
+                report.startswith('Boost.Test framework internal error: ') or
+                report.startswith('Test setup error: ')):
+            # boost.test doesn't do XML output on fatal-enough errors.
+            failure = BoostTestFailure('unknown location', 0, report)
+            return [failure]
+
         results = self._parse_log(log=log)
         shutil.rmtree(temp_dir)
         if results:
